apps/requirements.txt
azure-servicebus
azure-identity
aiohttp